+++ src/Octokit.Webhooks/Models/Organization.cs
@@ -10,7 +10,7 @@
         public string Login { get; init; } = null!;
 
         [JsonPropertyName("id")]
-        public int Id { get; init; }
+        public long Id { get; init; }
 
         [JsonPropertyName("node_id")]
         public string NodeId { get; init; } = null!;
