+++ src/Octokit.Webhooks/Models/SimplePullRequest.cs
@@ -11,7 +11,7 @@
         public string Url { get; init; } = null!;
 
         [JsonPropertyName("id")]
-        public int Id { get; init; }
+        public long Id { get; init; }
 
         [JsonPropertyName("node_id")]
         public string NodeId { get; init; } = null!;
@@ -29,7 +29,7 @@
         public string IssueUrl { get; init; } = null!;
 
         [JsonPropertyName("number")]
-        public int Number { get; init; }
+        public long Number { get; init; }
 
         [JsonPropertyName("state")]
         public SimplePullRequestState State { get; init; }
